+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Prefix your message with one of the following:
 - [Security] in case of vulnerabilities.
 -->
 
+## Unreleased
+
+- [Added] Generate
+  [dependabot](https://docs.github.com/en/free-pro-team@latest/github/administering-a-repository/configuration-options-for-dependency-updates)
+  configuration.
+
 ## v0.0.3 - 2020-11-03
 
 - [Changed] Remove unused `--repository` switch.