+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Prefix your message with one of the following:
 - [Added] Generate
   [dependabot](https://docs.github.com/en/free-pro-team@latest/github/administering-a-repository/configuration-options-for-dependency-updates)
   configuration.
+- [Changed] Add Postgres env vars so tests can run smoothly.
 
 ## v0.0.3 - 2020-11-03