+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Prefix your message with one of the following:
   [dependabot](https://docs.github.com/en/free-pro-team@latest/github/administering-a-repository/configuration-options-for-dependency-updates)
   configuration.
 - [Changed] Add Postgres env vars so tests can run smoothly.
+- [Added] Add helper method to create migration classes with `--active-record`.
 
 ## v0.0.3 - 2020-11-03