+++ CHANGELOG.md
@@ -14,10 +14,11 @@ Prefix your message with one of the following:
 ## Unreleased
 
 - [Added] Generate
-  [dependabot](https://docs.github.com/en/free-pro-team@latest/github/administering-a-repository/configuration-options-for-dependency-updates)
+  [dependabot](https://help.github.com/github/administering-a-repository/configuration-options-for-dependency-updates)
   configuration.
 - [Changed] Add Postgres env vars so tests can run smoothly.
 - [Added] Add helper method to create migration classes with `--active-record`.
+- [Added] Generate `.github/CODEOWNERS` template.
 
 ## v0.0.3 - 2020-11-03